clang: port clang indexer to new async API
This removes the "threaded" usage of IdeCodeIndexer and moves to
our new async/finish API.

It also creates a new CXIndex for each file, so we can more
reliably release memory after each parse. It probably means we
have to index some data more, but since this is a background
operation, I'd prefer to keep the memory usage lower at any
given time.