buildui: bind pipeline GListModel to stages list box
Now that pipeline is a GListModel, we can simplify our list of connected
build stages. This also ensures that things are updated after the pipeline
is initialized. Otherwise, we would show an empty list in the case that
the pipeline was not immediately setup fully. Such can happen when
installing dependent runtimes.