buildui: show build panel when build starts
I'm still not sure if we want to do this unconditionally in the long term,
but this is probably fine for now.

My concern long term is that we might get into a situation where the user
had another panel of importance open and keeps getting ping-pong'd away
from it.

But until our panel system is a bit more complex, this is likely a fine
solution.