code-index: limit main-loop index building to 1msec
To avoid stalling the main loop, this uses a GSource to add entries to the
database. It will process up to 1msec and then yield back to the loop.
Doing so allows higher priority items (such as the frame clock) to get
access to prepare/layout/etc instead of our indexing work.