unsaved-files: add API to reap old files
This removes the auto-reaping at initialization time (which would happen
since context would be set immediately) and instead adds API to force
a reaping.

Otherwise, we try to reap data dirs before we can build a full path to the
directory containing the buffers.

This fixes an error for which we reap all the old project data in the
.cache instead of just the buffers directory of the current project.