application-actions: Update and clean the copyright string
This builds on Michael’s tremendous and greatly appreciated work.
In the continued spirit I present a fix for the repeated issue of users
having to read ‘Copyright Copyright’ when reading … the copyright. With
this change that confusion and time waste will be a thing of the past,
literally, as it will also move Builder into the fresh new age of 2018.

AKA I’ve read Practical Typography so now I’m smart.