app: add helper to get translated projects directory
This implements the logic we want for getting the project directory. If the
gsettings "project-directory" has a value, and that directory exists, we
use that. If it doesn't exist, we check for "Projects" existing. If that
doesn't exist, we use the non-existant gsettings value (if it was
non-empty). Otherwise, we use the translated name "Projects" in the home
directory.

Fixes #341