code-indexer: pre-filter files by supported indexers
We want to generate our list of files to update on a worker
thread, but to do that accurately we need to know about what
indexers are supported. Accessing that data requires information
only accessible from the main thread.

To ensure we can stay threaded and perform that check safely,
we make a copy of the indexer information and proxy that to our
worker thread along with the other exclusive state.

This fixes #366