ctags: improve regex escaping
When consuming ctags files that use a regex to specify the location, we
have to convert the ctags regex into something GRegex can use. The regex
provided in the ctags file does not perform escaping on various features
like groups (), or repeats * and therefore we must do it for it.

Fixes #250