Add SPDX license identifier
The SPDX is an attempt at standardising the licensing attribution of
free and open source software in a way that is machine parseable.

Since we're shipping snippets that autogenerate the licensing blurb, it
would be a good thing to include this annotation in them, so that newly
written code can immediately be compliant.