file: add helper to peek at settings
If a file is open, we should have the file settings loaded. This allows
components to access the file settings without having to go through the
async flow, since it should be unnecessary. That simplifies some of our
code in completion providers if/when they need to use file settings.