subprocess: add a meson option to set PATH to run build commands
The previous commit unconditionally adds /usr/local/bin to PATH on
FreeBSD. It should work for most people, but the maintainer is still
required to patch the hard-coded /usr/local to the prefix chosen by
users because it is not possible to change the value after the build.

Since each operating system organizes its files in a different way, we
cannot expect a single PATH value to work on all operating systems.
This commit sets PATH for the four major BSD systems to the default
values used by their /bin/sh shell when PATH is unset with directories
including sbin in their names removed. It also add a meson option to
allow the package maintainer to set a different value without patching.