git: use array with bsearch() and apply drift in line callback
We need to adjust the line differences as we process through the hunk/line
changes (or deletions can be off).

Additionally, to simplify some of this code, we move from a GHashTable
containing line data (which has bad cache locality and lots of extra
allocations) to a cacheline friendly GArray with bsearch().

Fixes #386