config-manager: persist selected configuration
We want to reload the same configuration at startup when
possible. This still has a short period of time where we might
have a different config, but will re-apply the correct config
after provider loading has returned.

We then persist config changes to the GSettings (but we do not
watch the setting to avoid difficult cycles).