sysroot: add a new plugin allowing to specify the sysroot
A developer might want to specify the sysroot (the path to another system root where all the libraries will be used).
This is notably used when compiling for a foreign architecture or system.

This is ensured to work with the default GNOME project generated by GNOME Builder.

It is currently possible to specify the sysroot path and give extra arguments to Pkg-Config as
the differents multiarch implementations are difficult to handle.