flatpak: use pipeline runtime to specify build architecture
The runtime specifies an architecture and we should propagate that to the
build commands that flatpak uses. Doing so allows us to build for other
architectures like aarch64 (so long as you have an aarch64 device to
select and aarch64 flatpak runtime installed and properly setup static
linked qemu w/ aarch64 suport).