runtime-provider: specify runtime in bootstrap API
We might need to get an alternate architecture based on the device, so we
need to pass back the runtime instead of relying on the configuration.

Additionally, we need access to the selected device (which is going to be
hoisted out of the configuration so it can change).

We might want to consider a better abstraction for this going further, by
either passing the IdeBuildPipeline, or some sort of "IdeBuildSite" with
information about the build setup request.