mingw: remove mingw plugin
This device abstraction has not fit in well ever since we started on
runtimes, so it needs to be done differently in the future.

In particular, we need to ensure that we have a way to setup a toolchain
for a given project, and different build systems do that differently.
Meson, for example, will want a cross-build file describing the toolchain.