pipeline: simplify bootstrap process
Instead of passing multiple parameters around, we just pass the pipeline
now, for which components can get the data they're interested in.

To ensure that we have arch available early, we now get the device info
when invalidating the pipeline. This ensures that the runtime manager can
give proper arch info to the providers (who might translate that to get
the right flatpak runtime, etc).