qemu: add device provider for cross-architecture emulation
If qemu-user-static is installed, and we detect support from binfmt to run
non-native binaries, we can allow the user to select an alternate
architecture to do a native cross-build. This means that you're building
for another architecture, but instead of a cross-compiler, you're using
a native compiler runding under CPU emulation.

Not all build-systems/runtimes support this, and in the future we'll need
to improve the situation by warning the user about non-sensical setups.

However, if you're using flatpak, this allows us to install an aarch64 or
arm runtime and build against it (and run the result).

For the arm/flatpak case, you'll need to use 3.26 since the master SDK is
not available on the updated sdk build hosts. *womp womp*.