omnibar: provide visual feedback about build success/failure
Currently, you often have to peek into the omnibar to get the status of
the build. This gives us just a bit more feedback as to what happened with
a subtle background color on the fake entry box.

It would be nice to get some feedback from designers here, into better
ways to give this feedback that is not so dependent on color.