terminal: add new "Runtime" terminal option
This creates a new terminal that can be used to execute the program similar
to the environment used to run the application.

This is useful when you want more flexibility than we can give you with
the run button. Such as running unit tests that require a $DISPLAY, or
setting environment variables for the app.