todo: ignore additional patterns
We can ignore some more files that we know we won't want to look at. This
ensures we don't even read the content from disk or process it when we
know we'll just throw it away anyway.

Improves #438 although I'm confident there are other issues relevant to
fix for that codebase.