build-manager: avoid rediagnosing in a number of situations
This used to be a bit of an asynchronous build loop in certain situations,
such as when the pipeline needs to be advanced to get things like build
flags (and then re-forcing itself to advance immediately).

So now we only rediagnose if we haven't configured before, or if the
build request was for a real build.