code-index: port index builder to IdeTask
We have more control over where things are freed using IdeTask,
and we can use this object as a way to catch any issues that
might still exist before we push things through the rest of the
code base.

That said, I've used IdeTask to catch lots of bugs already, so
I'm pretty confident in it's usage now.