flatpak: fix wrong cwd when running a command through "flatpak build"
"flatpak build" will set the cwd of the run command to the build
dir, regardless of what directory it is run from.
Previously, this would cause us to ignore the cwd set on subprocess
launchers created in a flatpak runtime.
In turn, this would break subprocesses that rely on the cwd to
find configuration, such as eslint, which uses the cwd to resolve
the location of the source file and from there the .eslintrc.yml
file.