pipeline: incrementally load addins from idle callback
To reduce our chances of stalling the main loop, perform the pipeline
setup from an idle callback with lower priority than the frame clock.

We still get a stall on initial startup, which to fix we probably need
to make all this asynchronous. But combined with previous commits, it does
reduce the amount we stall in reaction to VCS changes.