source-view: disable snippets when performing undo
Unfortunately, we don't have a great way to deal with buffer changes while
snippet editing is in play. So we can just disable them when we come
across an undo request.

The problem at hand is tracking virtual edit-points across buffer changes
that are outside the snippet driver's control. This sort of "forms editing"
really needs to be handled as part of a core textview design if we want to
support this well.