buildconfig: add builddir option similar to flatpak manifests
This lets us specify either builddir = true or builddir = false in the
config group to indicate if an out-of-tree build directory should be used.

This is mostly useful for situations where the project cannot be built
out-of-tree.

We may want to alter subprocess launchers to take this into account when
running programs, but it should mostly happen automatically for pipeline
operations where builddir will now be within the project dir.