symbol-resolver: make find_references_async() transfer-full
This ensures that we can implement transfer properties through interface
virtual methods for which language bindings can fulfill.

We continue to use a transfer-container through the async tasks when in C
so that we don't leak if the caller fails to call the _finish() function of
the async pair.

When returning from the vfunc _finish, we clear the free func which
transforms the GPtrArray from transfer-container to transfer-full. This
means we also need to set the free_func in the consumer so that we can
continue on using g_autoptr(GPtrArray) as we want in the C code.