threading: backport IdeTask and IdeThreadPool updates
This allows us to simplify backports of other performance improvements by
giving us access to the new threadpool and tasking engine used on master
(which landed right after 3.28). It's proved to be stable and will help
us with some safety guarantees too if we choose to backport the GTask
conversion later on.