clang: stub out gnome-builder-clang client
This starts on a client that we can use to call RPCs on the new
gnome-builder-clang daemon. This manages spawning the process and
setting up the rpc client to communicate with.

Adding the rest of the API is mostly around marshaling the data. However,
in some cases we'll want to avoid unserializing in this API because we want
to have a single memory allocation that can be used by the plugin API
without creating lots of small strings.