clang: port clang highlighter to use IdeClangClient
This allows us to proxy the highlight request to the subprocess. Another
improvement is that we can try to update the document using our cached
highlight index, but then re-scan that area once we get the updated index
from the IPC peer.

Since we don't update the "invalidation position" in that case, we only
process the current line (where the user typed) with the old index. That
means we don't do a whole lot of duplicate work and also allow the user
to get immediate feedback for common symbols.