clang: apply unsaved buffers when calling clang API
We still need to plumb the buffers to the subprocess, but this
applies the unsaved buffers to the clang_parseTranslationUnit2
calls.

We need to take a reference from the main thread, so that we can
own that reference/copy in the worker thread in case the peer
updates the buffer while we're processing.