clang: introduce delay when updating highlight index
This is an expensive operation, so we want to reduce how often we do it.
In this change, we introduce a ~3 second delay to do the follow up update
using the resulting index (after highlighting with the current index).

That should only really be noticable by users typing a new symbol (like a
enum value) until the follow-up highlight occurs.