clang: start refactoring completion for clang
This is not exactly efficient yet, but it re-arranges things in a way that
will make it easier for us to port to the new completion engine when that
lands.

We should try to make this faster before we land the branch, because it
will probably be a week or two while we finish up the new completion
engine (which should render most of the runtime work unnecesary).