ctags: make ctags search more lazy
This now performs the search in a thread and avoids creating an GObjects
until they've been requested by the GListModel interface. That means our
dynamic allocations are now just an array to point to the index entry (and
it's priority for sorting). The objects can now be discarded at will when
they are no longer visible.