todo: aggressively ignore more data from todo
The todo module has an optimization to avoid lots of small strings being
allocated. In doing so, we potentially pull in data that we cannot know
we need to ignore until after it has been delivered from the subprocess.

This tries to ignore node_modules, a commonly used directory for npm
dependencies from even beeing scanned by grep.

Related to #509