snippet: reduce parsing and allocations
This reduces the chances that we make allocations by only using the
info pointer in the completion proposal, and then parsing the snippet
on expansion.

Also, by moving the get_comment() to the provider, we can avoid having to
subclass proposals with so much code.

I think we might want to just use G_TYPE_OBJECT for proposals.