completion: allow tweaking priority for completion
Based on the completion location, we may want to alter the priority of the
group of completion items. This allows the snippet provider to be lower
priority when it isn't at the beginning of a line, so that more interesting
results are at the top.

Another thing we can do, is to create a special internal list model that
is used for very likely results and let providers register them with the
context.