words: add words completion plugin
This transforms our previous word completion plugin to use the new
plugin engine and background scanning.

We still need to make this only display when the user requests
activation so that it isn't over active.

It also works around a GtkSourceView bug for the time being, which
we can remove once it has been addressed.