snippet: add revamped snippet engine
When writing this, there was some anticipation that it would land upstream
in GSV, but I don't think that will happen given ABI stability and what
not. So we might as well save us some character columns.

More importantly, this adds the ability to avoid parsing all the snippet
content up front (and instead delay that until they are used). We still
need to duplicate some strings for metadata, but a couple hundred objects
can be avoided now.

This doesn't enable the new snippet engine, only lands it so that it can
be used from other API as we merge the new completion engine. The existing
code will be removed in a future commit.