words: port word completion to new completion engine
This moves the word completion provider to the new storage engine. It only
shows results when the user has requested completion (ctrl+space) since
otherwise we'd be doing an awful lot of unnecessary buffer scanning.

We may change that in the future to just do delayed buffer scanning unless
requested by the user.