completion: prefer resize to moving X
This will allow us to keep the window alignment and shrink the width which
is more often than not what we want.

It does have a downside, in that if we're near the edge we might have an
issue not being visible. But it appears that move_to_rect doesn't really
allow us to prefer resizing X before sliding X.