keybindings: Add Sublime Text keybindings
Adds a setting so that the option appears in the Preferences pane.

These bindings have a few idiosyncracies which are listed at the bottom
of the keybinding CSS file in a comment. In particular, to emulate
Sublime Text nicely, we should make a few keybindings do something
different depending on whether any text is selected.