plugins: use dzl_clear_pointer()
This uses the new dzl_clear_pointer() to free data instead of
g_clear_pointer() because it allows us better type checking of
the callback parameter and the data to be freed.

We want to keep using this going forward so that we catch errors
sooner from miss-use of incorrect callbacks.