cpack: add completion provider for header includes
This will try to locate headers that are available in the
include paths for the file (based on build flags).

It activates when in the "path" context for a C/C++ file.

That is sort of dodgy still if you have a line line:

  #include <|

But works fine if you have the trailing > like:

  #include <|>

It's less of an issue with "" includes since we insert a
matching " by default.