todo: tell git-grep to avoid long lines
This allows us to ignore long lines while git is processing the directory
tree in the subprocess. Doing so means we don't pollute our GBytes in the
UI process with data that will never be displayed to the user. Especially
when said data is pathological minified JS in a given project.

It is important that we do the --and -e length check *after* the TODO
check so that grep doesn't have to look at every line length as that is
quite expensive.