todo: delay loading until user selects todo panel
We can avoid all subprocess grep work until the user actually chooses to
look at TODO items in the panel. This does exactly that. Some larger
projects like WebKit can take a number of seconds to complete loading. So
we introduce an empty-state page so that the user knows background work
is going on while we locate todo items.