wip: start on hover abstraction for sourceview
The IdeHoverProvider and IdeHoverContext abstractions are used as a
means to communicate information to be displayed in a popover above
the IdeSourceView similar to that of a tooltip.

However, the popovers can stick around for long enough for the user to
enter them and manipulate information within the popover. When the
pointer leaves or goes out of the grace area, the popover is
immediately dismissed similar to a tooltip.