symbol-tree: add hover provider to display symbol info
This adds a new hover provider that will display the name of the
symbol for the type being hovered. It also provides links to the
definition, declaration, and canonical location if provided by
the symbol resolver.

The user can click on the link to be taken to directly to the
source location.