flatpak: ensure widget access after task propagation
Normally, we dispose of the source_object and task_data when the
IdeTask propagates the return value. This helps ensure that we
avoid leaks and protect against objects finalized in the wrong
thread. However, we need access to the widget in this case so
we need to keep the source_object around.

We still get the protection from the threaded worker sending the
result back with the task reference.

Fixes #560