hover: dismiss hover popover based on crossing event location
If we detect that the crossing event is into a child window, we don't want
to dismiss the popover (it could be a button, webview, etc). However, if
we are getting a crossing event that is leaving the popover, we want to
dismiss the hover popover.

This should help in situations where leaving the hover popover into a
different widget (like the sidebar) left the popover visible.