flatpak: drop legacy repo cleanup
A long time ago we had a repository that would be used for installations.
That was only for a single release of Builder. It's been a few releases
and so we can drop this now.

It also allows us to shave a bit of time off startup while we synchronously
block for the flatpak command.