app: make application startup feel faster when loading files
We can cheat and display the window sooner while the project loads so that
the user thinks things are loading faster. It's not really, we still take
about 1 second to be useful, but we shave about 400msec off the time to
first window display when used like "gnome-builder foo.c".

Potentially we can do some background loading on a thread while the window
resources are created, but that doesn't seem to be all that much savings
currently.