CI: Workaround cache misses.
Gitlab CI by default zips it's cache target which causes the files
to lose their xattributes that ostree relies upon.

This workaround makes a tarball of the cache target first, which is then
zipped by the Gitlab CI. Then on before_sciprt it extracts it in place.

Hopefully tarballs will be the default behavior in the Future™ and this
commit will get reverted.

Relevant issue: https://gitlab.com/gitlab-org/gitlab-ce/issues/37444