Allow packagers to specify the Python libprefix.
When packaging for Nix, the directory that pygobject finds its overrides
in is not writable. It'll usually look something like this, and can only
contain files built by the pygobject package itself:

/nix/store/58qam3zgdcvvsz4g081pp98kg9i5v7rb-python3.6-pygobject-3.26.1/lib/python3.6/site-packages/gi/overrides

This makes Builder's pygobject_override_dir heuristic fail, because
Builder's libdir is not a prefix of pygobject's libdir.

Fortunately, what Builder actually needs is for its overrides to be on
the Python import path when it runs, and that's easier to arrange. We
just need to be sure that Ide.py gets installed to

<libdir>/python<version>/site-packages/gi/overrides

So this patch allows packagers to optionally specify the
"python<version>" portion of the path, and just constructs the rest of
the path using Builder's libdir.

The existing auto-detection code is still used unless the
-Dpython_libprefix option is explicitly specified to override it.