flatpak: allow applications to talk to portals
Portals are explicitly designed to give sandboxed applications access to
resources in a secure manner, and flatpak therefore grants all applications
permission to talk to any portal.

As a result, few manifests will explicitly request that permission, so
to have apps work as expected when run through our flatpak-runner instead
of flatpak itself, we should grant them the same permission.