git: initialize git submodules from build pipeline
We can download submodules from the IDE_BUILD_PHASE_DOWNLOADS if we detect
that any are missing from `git submodule status` as lines prefixed with -.

We only run this once, to avoid being overly annoying when extracting build
flags.

This has brought to my attention that we probably also need a dependency
updater addin for submodules.