git: add dependency updater for git submodules
When clicking "Update Dependencies" in the UI, we can now update git
submodules along with the flatpak dependencies.

In the future, we probably want to remove the build pipeline advancement
from the updaters themselves, and instead do that from the code that
creates the updater plugin instances so that we don't race.