buffers: create the directory tree before saving
In case the user moved the directory out from under Builder, we should
recreate all of the directories needed to be able to save the file.

It would be nice to make this async in the future, but there is not
currently a variant of this API that is async.

Fixes #659