buffer: set buffer modified when backing file is lost
If we detect that the backing file has been lost (failed to locate
GFileInfo) then ensure that we mark the buffer modified so that the user
can save it again.

I might rather see this done in a way that tells the user the file was
deleted, similar to the "reload" bar we have. But instead show two options
to close/save.

This at least allows them to save the file though.

Fixes #659