sysmon: remove sysmon plugin
We already have CPU graphs as part of our profiler based on Sysprof. We
don't really need the additional overhead of this, and that can be pretty
great on Wayland with cairo_image_surface_t due to the texture uploads.
Especially at 60fps which we can usually hit.

Instead, we'll just rely on the profiler and add additional probes there
for the other features we care about long term (mem/net/disk).