rls: prepend rustup path to PATH
The existing code replaces PATH altogether. But this doesn't work when
rustc ends up calling out to utilities like cc. The end result is that
rls silently stops working. Instead, we will retain the existing path
and merely add the rustup location to it. We put the rustup location
first because this will most likely be the most up to date rust
toolchain.

Fixes: #669