glade: start on basic glade plugin for gtk3 support
I'm not sure the lifetime of the glade plugin, because we know that
we'll have to write something new for gtk4. But we have a couple of more
releases with gtk3, so we can use this for now.

We still need to add the signal panel, and do our modified version of
the editor panel. We also need a couple accessor widgets.

This is just a minimal prototype for the mechanics.