glade: add basic support for working with glade documents
I'm not sure the lifetime of the glade plugin, because we know that
we'll have to write something new for gtk4. But we have a couple of more
releases with gtk3, so we can use this for now.

There is still much more to do, but this is basic enough that it works
to a degree that is useful.

Some of the missing parts include:

 * Better visibility of what is going on
 * Toolbar for various palette widgets
 * Hooking up signals to source code
 * Missing icons which are in the works

and probably a bunch more.