completion: improve fuzzy matching
This ensures that we match 'E' before 'e' in 'Endianness' if the input
character is 'e'. Otherwise, 'end' won't match and that is not the
expected behavior.

This just takes the earliest of 'E' or 'e' as the match for any character
that has a toupper() variant, at the cost of running additional matches.

This code does need to stay fast, but it seemed fast enough in testing.

Fixes #673