buildui: move and refactor buildui into plugins
This rather large commit moves the buildui components into the plugins
directory. It always was a plugin, but previously was embedded into
libide which is no longer necessary.

Also, the build preferences has been abstracted into using addins so that
plugins can extend what can be toggled for a given build configuration.