code: add new libide-code static library
This new library depends on libide-core, libide-io, and others to create
a static library containing utilities and types for managing code. It also
includes the buffer manager since that is so intertwined with the other
components contained herein.

Symbols, Diagnostics, Trees, Locations, and Ranges have all been converted
into GObject so that we have an easier time binding things in the future.