libide-gui: add new libide-gui static library
This adds the new refactored UI code in the form of libide-gui. Some of
the main changes to the design include the abstraction of IdeWorkbench
from a GtkWindow to a GtkWindowGroup. It now contains multiple workbench
IdeWorkspace (GtkWindow). This allows for multi-monitor setups so that
users may have multiple views of code and content.

The IdeContext creation also no longer needs to have a project to load
initially. This allows for a non-project mode while still building on
the abstractions provided by IdeContext/IdeObject.

However, plugins should check to see if they're in a project mode and
avoid accessing foundry components in that case.