libide-greeter: add new libide-greeter static library
This creates a new static library that contains the greeter. It breaks
the greeter into a separate window so that we can vastly simplify the
code in Builder with regards to workspaces.

Some code is moved to a plugin so that it can hook into command line
options and provide the new -g option to jump to the greeter.