libide-editor: add libide-editor static library
This creates a new libide-editor static library and moves a number of
components into a separate plugin. The plugin was previous included
directly in libide, and is no longer the case (now that there is no more
libide shared library).

Various components have been renamed to follow our naming semantics such
as Page, Pane, and Panel.