command-bar: port command bar to libide-editor
This also implements a new design as part of this work to make the
command bar front-and-center when used. We have lost a little bit of
features in the process (naming the tab completion still needs work as
well as dyanmic data such as %s/foo/bar).

We expect more changes in this plugin in the future as design work
fleshes out more of a command bar experience.