git: port to new libide designs
This ports the git plugin to use new libide designs. It also adjusts the
prefix for the plugin to Gbp from Ide (which was holdoff from the original
libide days).

Some components have been simplified, such as the buffer change monitor
integration now provided via the IdeBufferAddin extension.

The new index monitor will also allow us to be a bit more clever in the
future about tracking changes to activate branches and commits. However,
it does not do that yet.