projects: propagate build-system-hint when opening from dialog
If we know the build system to use up-front, such as from the file-chooser
dialog with filter set, then we can avoid having that overridden by build
system discovery.

This also adds that hint to the persistent recent-info so that any future
opening of the project re-selects the same build system plugin.

Fixes #753