flatpak: raise priority of build-system discovery
This should be higher than the build systems priority because the project
is explicitly telling us what it should be.

Projects opened with the file-chooser will have the plugin that provided
the file-filter pinned when opening the project. That will get saved to
the recent-info so that followup loads get the same plugin, so having a
higher priority than the build systems themselves is fine.