code: short-circuit save when we think there are no changes
If we have not yet gotten notification of the backing file changed, and
the buffer thinks there are no local modifications, then we can skip the
save operation.

This also helps a situation where we modify a .json flatpak manifest and
invalidate the build pipelines immediately due to the inotify event.

Fixes #757