foundry: rename IdeBuildPipeline and IdeBuildStage
This renames IdeBuildPipeline to IdePipeline and IdeBuildStage to
IdePipelineStage. Additionally, IdeBuildPhase is now IdePipelinePhase.

This should help in contributors understanding the difference between
the IdeBuildSystem components and the production pipeline and it's
components.

I didn't try very hard to keep whitespace nice, because we'll probably
do a large round of that with automation tools later and I'd rather do
that as a single pass.

Also, I renamed the execute APIs to build since execute didn't really
make sense in the concepts of a build pipeline that can
build/rebuild/clean.

I would not be surprised if there is a bit of fallout from this, so better
to get it in now than later.