terminal: allow opening filename:line:column
We try to discover the file from the filename using the resovle_file API
from the workbench. That is somewhat difficult because we don't have the
active CWD from the point when the line was printed. So we do a bit of
discovery in case we don't find it.

We will want to extend this handle some other filename variants such as
without line or columns. But this is a start.