editor: try to reuse existing workbenches
If we discover a workbench that satisfies the necessary working directory
for the file(s) we are opening, then reuse it instead of creating a new
workspace window.

Also, we don't try to separate into multiple workbenches here, but the
open API in IdeWorkbench will do that for us (should multiple files be
selected).

Fixes #775