git: begin switch_branch implementation
To do this well, we will need lots of fallbacks to handle situations where
the contents are different. For now, we'll just set SAFE. However, in the
future we should provide the user some choices in UI to browse the changes
and select what to do. That should fit fine into the async/finish model
since we can just delay the operation with a modal dialog.