code-index: add GbpCodeIndexPlan
This object can be used to help discover the plan we will take when
building our code indexes. Future code will expand on this so that we
can extract bits from the IdeCodeIndexBuilder implementation.

We should also be able to provide progress support going forward as we
will know what needs to be done. I'm also interested in seeing a plan
executor object (which is where pausability can be implemented).