editor: allow keeping search box visible on view focus-in
We try to only keep the search context alive while the bar is visible. To
allow for the results to stick around, we can allow keeping the search bar
visible by clicking somewhere in the document.

To allow for this, we need to clear the saved bit in the source view so
that returning to the source view does not cause focus-in-event tracking
to restore the previous cursor position.

Fixes #792